+++ packages/taro-components-rn/TCRNExample/example/EXWebView.js
@@ -1,0 +1,10 @@
+import React, { Component } from 'react'
+import { WebView } from '../../dist'
+
+export default class EXSlider extends Component {
+  render () {
+    return (
+      <WebView src="https://taro.aotu.io/" style={{ width: 300, height: 150 }} />
+    )
+  }
+}