+++ packages/taro-components-rn/gulpfile.js
@@ -4,6 +4,7 @@ const babel = require('gulp-babel')
 const ts = require('gulp-typescript')
 
 const tsProject = ts.createProject('tsconfig.json')
+const distPath = 'dist'
 
 const compatibleJsFiles = 'src/**/*.js'
 const staticFiles = 'src/**/*.png'
@@ -16,26 +17,27 @@ function scripts () {
     .pipe(babel({
       presets: ['module:metro-react-native-babel-preset']
     }))
-    .pipe(dest('dist'))
+    .pipe(dest(distPath))
 }
 
 function typescripts () {
   return tsProject
     .src()
     .pipe(tsProject())
+    .on('error', (err) => {
+      console.log(err)
+    })
     .js.pipe(babel({
       presets: ['module:metro-react-native-babel-preset']
-    })).pipe(dest('dist'))
+    })).pipe(dest(distPath))
 }
 
 function images () {
   return src([staticFiles])
-    .pipe(dest('dist'))
+    .pipe(dest(distPath))
 }
 
 // watch(compatibleJsFiles, scripts)
-
 // watch(staticFiles, images)
 
 exports.default = parallel(scripts, typescripts, images)
-// exports.default = typescripts