+++ packages/taro-h5/src/api/privateApis/onPullDownRefresh.js
@@ -1,4 +1,0 @@
-
-export const onPullDownRefresh = (opt) => {
-
-}